Change the background color depending on value of GITC
GITC is an environment variable set in `.dotfiles/.bash/prompt.sh` to
determine the color of the git branch in the PS1. When using Solarized
Light in the shell, the white color is nearly invisible, so I set it to
black instead. This change allows vim to alter its background
accordingly.

Also, remove some outdated, commented-out settings