OK, let's fix tmux (and paste in Vim) for real, no auto-saving tmux
No longer having continuum auto back-up. This is because the subshell
required for vim's copy and paste to clipboard work screws up the way
that resurrect saves the currently running programs.

This is a tradeoff I am willing to accept for now, especially because
my development environment does not often change.